Accessing contact info update
Modified view contact and edit contact activities to pull contact attributes from database instead of intent.

Modified getItemID(String number) in Database helper and created getContact(int ID) to support this.

Created a new constructor in Contact class so that img Uris can be passed as a String or Uri.